refactor: extract Bytes codec out of Misc
Move the Bytes <-> Double/String/hex codec (btsToNum, numToBts,
strToBts, bytesToBts, btsToStr, btsToUnescapedStr and their internal
hex/Word8 helpers) from the Misc grab-bag into a dedicated Bytes
module. Consumers import Bytes directly; Misc sheds the bytestring,
binary, bits and numeric imports it only needed for the codec.

First of several steps splitting Misc into focused modules.
Behaviour-neutral: -Werror build clean, all 1078 tests pass.
