+++ src/CST.hs
@@ -11,6 +11,7 @@
 module CST where
 
 import AST
+import Bytes (btsToNum, btsToStr)
 import qualified Data.Text as T
 import Misc
 import qualified Yaml as Y
